refactor: extract close region helper
